NSItemProvider for EKReminder
By providing a NSItemProviderWriting extension and a DropDelegate we can
add support for dragging an EKReminder onto an EKCalendar to move items
between calendars. In the future we will probably want to use more
structured data, but for now we will use NSUrl. By encapuslating it in
our own extension, we should be able to change the logic in our library
without our application needing to care about the details.